Fix all source code, Update in 2018-6-12
