+++ Test_Binary/Beta-Vesion3.1/[Note Information] EODF CUI_Program file crypto function(Version 3.1).txt
@@ -6,7 +6,7 @@ Chinese:
 如果，你们想要我开发这个程序的Linux版本 。
 请尽量可能的帮助我的开发，请在我的github网站空间上留言。谢谢你们的支持
 并且，程序从Beta3.0版本，替换升级到测试的 Beta 3.1版本的变化比较大。
-作者本人(Twilight-Dream-Magic)已经试验成功，新的加密和解密的核心算法，对于单字节文件流数据和密钥之间，来进行处理计算，程序变得比较有用和表现很好。
+作者本人(Twilight-Dream-Of-Magic)已经试验成功，新的加密和解密的核心算法，对于单字节文件流数据和密钥之间，来进行处理计算，程序变得比较有用和表现很好。
 更新后，被保护的文件的安全性提高一些，被处理文件的密文数据信息更多了。不过，文件的处理速度还是很慢。
 
 English:
